feat: endurecer caja, compras/unidades e instalador IIS Windows
Incluye integridad de sesión de caja, moneda/unidades en compras, referencia USD en POS, permisos sensibles y corrección del handler PHP del instalador IIS (appcmd) para despliegues offline.

+++ database/migrations/2026_01_17_003732_create_purchase_details.php
@@ -4,14 +4,15 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     public function up(): void
     {
         Schema::create('purchase_details', function (Blueprint $table) {
             $table->id();
             $table->foreignId('purchase_id')->constrained()->cascadeOnDelete();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
-            $table->integer('quantity');
+            $table->decimal('quantity', 14, 4);
             $table->decimal('price', 10, 2);
             $table->decimal('subtotal', 10, 2);
         });
